Remove ConfigParser class and move generateMap method to ConfigUtils.
